+++ doc.go
@@ -35,7 +35,7 @@ Source code and other details for the project are available at GitHub:
 
 Current Version
 
-8.5.0
+8.5.1
 
 Installation